requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

# Deep Learning & Computer Vision
torch==2.1.2+cu118; sys_platform == 'linux'
torch==2.1.2+cu118; sys_platform == 'win32'
torchvision==0.16.2+cu118; sys_platform == 'linux'
torchvision==0.16.2+cu118; sys_platform == 'win32'
ultralytics==8.1.0
paddlepaddle-gpu==2.6.0
paddleocr==2.7.0.3
opencv-python==4.9.0.80

# Database & Async
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Utility
pyyaml==6.0.1
pydantic==2.5.3
loguru==0.7.2
psutil==5.9.8
GPUtil==1.4.0
python-Levenshtein==0.25.1
